backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
joblib==1.3.2
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==3.0.0
torch==2.0.1
huggingface-hub==0.21.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
